+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added a file for baking key concepts called (`bake_chapter_key_concepts`) and the respective specs.
 (minor change) only affects the book if called from the recipe
 * Changed the main gem source file to have the same name as the gem (`openstax_kitchen`) so that you don't have to `require` a different name than you use in your `gem` call.
+* Added a file for baking composite chapters called (`bake_composite_chapters`) and the respective spec.
+(minor change)
 
 ## [1.0.0] - 2020-12-15